add 2d polar plot
