+++ src/plopp/graphics/figure.py
@@ -46,9 +46,19 @@ def figure1d(*args, style: Literal['line'] = 'line', **kwargs):
         return backends.figure1d(LineView, *args, **kwargs)
 
     if style == 'polar':
-        from .polarview import PolarView
+        from .lineview import LineView
 
-        return backends.figure1d(PolarView, *args, **kwargs)
+        return backends.figure1d(
+            LineView,
+            *args,
+            **{
+                **kwargs,
+                **{
+                    'artist_maker': backends.polar_line,
+                    'canvas_maker': backends.polar_canvas,
+                },
+            },
+        )
 
     raise ValueError(f'Unsupported style={style} for figure1d.')
 
@@ -83,6 +93,21 @@ def figure2d(*args, style: Literal['image'] = 'image', **kwargs):
         from .imageview import ImageView
 
         return backends.figure2d(ImageView, *args, **kwargs)
+
+    if style == 'polar':
+        from .imageview import ImageView
+
+        return backends.figure2d(
+            ImageView,
+            *args,
+            **{
+                **kwargs,
+                **{
+                    'artist_maker': backends.polar_image,
+                    'canvas_maker': backends.polar_canvas,
+                },
+            },
+        )
 
     raise ValueError(f'Unsupported style={style} for figure2d.')
