+++ src/plopp/graphics/lineview.py
@@ -7,6 +7,7 @@ import scipp as sc
 
 from .. import backends
 from ..core import View
+from ..core.typing import CanvasLike
 from ..core.utils import make_compatible, name_with_unit
 
 
@@ -80,8 +81,8 @@ class LineView(View):
         figsize: Tuple[float, float] = None,
         format: Optional[Literal['svg', 'png']] = None,
         legend: Union[bool, Tuple[float, float]] = True,
-        artist_maker=backends.line,
-        canvas_maker=backends.canvas2d,
+        artist_maker: Optional[View] = None,
+        canvas_maker: Optional[CanvasLike] = None,
         **kwargs,
     ):
         super().__init__(*nodes)
@@ -91,7 +92,9 @@ class LineView(View):
         self._mask_color = mask_color
         self._kwargs = kwargs
         self._repr_format = format
-        self._artist_maker = artist_maker
+        self._artist_maker = backends.line if artist_maker is None else artist_maker
+        if canvas_maker is None:
+            canvas_maker = backends.canvas2d
         self.canvas = canvas_maker(
             cbar=False,
             aspect=aspect,
